Refactor: Moved TextFile and BinaryFile implementations to cpp file
